+++ sknetwork/data/models.py
@@ -57,7 +57,7 @@ def block_model(sizes: Iterable, p_in: Union[float, list, np.ndarray] = .2, p_ou
     np.random.seed(random_state)
     sizes = np.array(sizes)
 
-    if type(p_in) == float:
+    if isinstance(p_in, (np.floating, float)):
         p_in = p_in * np.ones_like(sizes)
     else:
         p_in = np.array(p_in)