+++ sknetwork/utils/__init__.py
@@ -5,10 +5,10 @@ from sknetwork.utils.format import *
 from sknetwork.utils.kmeans import KMeansDense
 from sknetwork.utils.knn import KNNDense, CNNDense
 from sknetwork.utils.membership import membership_matrix
+from sknetwork.utils.neighbors import get_neighbors
 from sknetwork.utils.parse import edgelist2adjacency, edgelist2biadjacency
 from sknetwork.utils.simplex import projection_simplex, projection_simplex_array, projection_simplex_csr
 from sknetwork.utils.ward import WardDense
-
 
 
 class Bunch(dict):
@@ -16,19 +16,18 @@ class Bunch(dict):
     Dictionary-like object that exposes its keys as attributes.
 
     This code is taken from scikit-learn.
-    >>> b = Bunch(a=1, b=2)
-    >>> b['b']
-    2
-    >>> b.b
-    2
-    >>> b.a = 3
-    >>> b['a']
+    >>> bunch = Bunch(a=1, b=2)
+    >>> bunch['a']
+    1
+    >>> bunch.a
+    1
+    >>> bunch.b = 3
+    >>> bunch['b']
     3
-    >>> b.c = 6
-    >>> b['c']
-    6
+    >>> bunch.c = 4
+    >>> bunch['c']
+    4
     """
-
     def __init__(self, **kwargs):
         super().__init__(kwargs)